INtroducing the tonemapping job in its separate file and clean DeferredLightingEffect
