+++ libraries/render-utils/src/ToneMappingEffect.h
@@ -1,0 +1,46 @@
+//
+//  ToneMappingEffect.h
+//  libraries/render-utils/src
+//
+//  Created by Sam Gateau on 12/7/2015.
+//  Copyright 2015 High Fidelity, Inc.
+//
+//  Distributed under the Apache License, Version 2.0.
+//  See the accompanying file LICENSE or http://www.apache.org/licenses/LICENSE-2.0.html
+//
+
+#ifndef hifi_ToneMappingEffect_h
+#define hifi_ToneMappingEffect_h
+
+#include <DependencyManager.h>
+#include <NumericalConstants.h>
+
+#include <gpu/Resource.h>
+#include <gpu/Pipeline.h>
+
+class RenderArgs;
+
+class ToneMappingEffect {
+public:
+    ToneMappingEffect();
+    virtual ~ToneMappingEffect() {}
+
+    void render(RenderArgs* args);
+
+private:
+
+    gpu::PipelinePointer _blitLightBuffer;
+
+    // Class describing the uniform buffer with all the parameters common to the tone mapping shaders
+    class Parameters {
+    public:
+        
+        Parameters() {}
+    };
+    typedef gpu::BufferView UniformBufferView;
+    gpu::BufferView _parametersBuffer;
+
+    void init();
+};
+
+#endif // hifi_ToneMappingEffect_h
